+++ lib/SKGB/Intern.pm
@@ -5,7 +5,7 @@ use Mojo::Log;
 use SemVer;
 
 #our $VERSION = Perl::Version->new( '2.0.0_5' );
-our $VERSION = SemVer->new( '2.0.0-a15' );
+our $VERSION = SemVer->new( '2.0.0-a16' );
 
 
 sub startup {
@@ -21,7 +21,6 @@ sub startup {
 	push @{$app->commands->namespaces}, 'SKGB::Intern::Command';
 	
 	$app->plugin('SKGB::Intern::Plugin::Neo4j');
-	$app->plugin('SKGB::Intern::Plugin::SessionManager');
 	$app->plugin('SKGB::Intern::Plugin::DigestAuth');
 	$app->plugin('SKGB::Intern::Plugin::AuthManager');